new added

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: praewpruettipongsapuk <praewpruettipong    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/11 10:29:06 by praewpruett       #+#    #+#             */
-/*   Updated: 2023/10/11 10:44:26 by praewpruett      ###   ########.fr       */
+/*   Updated: 2023/10/15 10:35:22 by praewpruett      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int		i;
+	size_t		i;
 
 	i = 0;
 	while (s[i])
